+++ egs/tedlium2/asr1/run.sh
@@ -217,16 +217,16 @@ if ${use_unsegmented}; then
     for task in ${recog_set}; do
 	task_new=${task}_unsegmented
 	if [ -d "${dir}/${task_new}" ]; then
-	    rm -r ${dir}/${task_new}
+	    rm -r ${dir:?}/${task_new}
 	fi
 	mkdir -p ${dir}/${task_new}/wavs
 	python local/conct_wav.py \
-	    ${dir}/${task}/wav.scp \
-	    ${dir}/${task}/segments \
-	    `pwd`/${dir}/${task_new}/wavs
+	    "${dir}/${task}/wav.scp" \
+	    "${dir}/${task}/segments" \
+	    "$(pwd)/${dir}/${task_new}/wavs"
 	cp ${dir}/${task_new}/wavs/wav.scp ${dir}/${task_new}/
-	cat ${dir}/${task}/text | python ./local/conct.py > ${dir}/${task_new}/text
-	cat ${dir}/${task_new}/text | cut -f 1 | awk {'print $1"\t"$1'} > ${dir}/${task_new}/utt2spk
+	python ./local/conct.py < ${dir}/${task}/text > ${dir}/${task_new}/text
+	awk '{print $1"\t"$1}' < ${dir}/${task_new}/text > ${dir}/${task_new}/utt2spk
 	cp ${dir}/${task_new}/utt2spk ${dir}/${task_new}/spk2utt
     done
